Moved Barista class to task3, to be together with all coffees. Made all coffee classes not public, except for barista class. Made Main in task4 import Barista.
